+++ doc/tutorial/search/views/minimiser/index.md
@@ -62,7 +62,7 @@ assignment!
 Task: Obtain the minimisers for "CCACGTCGACGGTT" with an ungapped shape of size 4 and a window size of 8.
 \endassignment
 \solution
-\include doc/tutorial/ranges/views/minimiser/minimiser_solution1.cpp
+\include doc/tutorial/search/views/minimiser/minimiser_solution1.cpp
 \endsolution
 
 If you have completed the assignment, you probably wonder what these large numbers mean. As explained above, the
@@ -71,7 +71,7 @@ XORs the hash values with a random seed (Default: 0x8F3F73B5CF1C9ADE). How would
 then?
 Well, you just use XOR again!
 
-\include doc/tutorial/ranges/views/minimiser/seed_example.cpp
+\include doc/tutorial/search/views/minimiser/seed_example.cpp
 
 From these hash values, you can obtain the sequence they are representing by transforming the numbers to base 4. (For
 example, 182 is "2312" in base four and therefore represents "GTCG".)
@@ -102,7 +102,7 @@ same hash value.
 
 \endassignment
 \solution
-\include doc/tutorial/ranges/views/minimiser/minimiser_solution2.cpp
+\include doc/tutorial/search/views/minimiser/minimiser_solution2.cpp
 \endsolution
 
 ### Ignoring the backward strand
@@ -112,7 +112,7 @@ the backward strand should not be considered. If this is the desired behaviour t
 to be used. Unlike the `seqan3::views::minimiser_hash`, `seqan3::views::minimiser` does not hash the values for you, so
 you have to do this yourself. But fear not, `seqan3::views::kmer_hash` makes this really easy for you!
 
-\snippet doc/tutorial/ranges/views/minimiser/minimiser_snippets.cpp minimiser
+\snippet doc/tutorial/search/views/minimiser/minimiser_snippets.cpp minimiser
 
 
 This syntax will result in minimisers with k-mer size 4 and a window-length of 8 (5 + 4 - 1). (So, to determine the
@@ -144,11 +144,11 @@ In order to ensure that this is the desired behaviour, using `seqan3::views::min
 Last but not least, `seqan3::views::kmer_hash` and `seqan3::views::minimiser` do not have a seed parameter. So, in order
 to obtain a random ordering, you have to XOR the view yourself. This can be done with the following command:
 
-\snippet doc/tutorial/ranges/views/minimiser/minimiser_snippets.cpp minimiser_seed
+\snippet doc/tutorial/search/views/minimiser/minimiser_snippets.cpp minimiser_seed
 
 \assignment{Assignment 3: Fun with minimisers III}
 Task: Repeat assignment 2 but this time do not consider the backward strand.
 \endassignment
 \solution
-\include doc/tutorial/ranges/views/minimiser/minimiser_solution3.cpp
+\include doc/tutorial/search/views/minimiser/minimiser_solution3.cpp
 \endsolution
