+++ doc/tutorial/search/views/minimiser/minimiser.png
