+++ doc/tutorial/search/views/minimiser/minimiser_snippets.cpp
@@ -1,9 +1,10 @@
 #include <iostream>
-#include <vector>
 #include <seqan3/std/ranges> // include all of the standard library's views
+#include <vector>
+
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/range/views/kmer_hash.hpp>
-#include <seqan3/range/views/minimiser.hpp>
+#include <seqan3/search/views/minimiser.hpp>
 
 using seqan3::operator""_dna4;
