+++ doc/tutorial/search/views/minimiser/minimiser_solution1.cpp
