+++ doc/tutorial/search/views/minimiser/minimiser_solution2.cpp
