+++ doc/tutorial/search/views/minimiser/minimiser_solution3.cpp
@@ -1,7 +1,7 @@
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/core/debug_stream.hpp>
 #include <seqan3/range/views/kmer_hash.hpp>
-#include <seqan3/range/views/minimiser.hpp>
+#include <seqan3/search/views/minimiser.hpp>
 
 using seqan3::operator""_dna4;
 using seqan3::operator""_shape;
