+++ doc/tutorial/search/views/minimiser/seed_example.cpp
