+++ include/seqan3/range/views/minimiser_hash.hpp
@@ -15,7 +15,7 @@
 #include <seqan3/core/detail/strong_type.hpp>
 #include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/kmer_hash.hpp>
-#include <seqan3/range/views/minimiser.hpp>
+#include <seqan3/search/views/minimiser.hpp>
 
 namespace seqan3
 {
