+++ test/snippet/search/views/minimiser.cpp
@@ -2,7 +2,7 @@
 #include <seqan3/core/debug_stream.hpp>
 #include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/kmer_hash.hpp>
-#include <seqan3/range/views/minimiser.hpp>
+#include <seqan3/search/views/minimiser.hpp>
 
 using seqan3::operator""_dna4;
 using seqan3::operator""_shape;
