+++ test/unit/range/views/CMakeLists.txt
@@ -32,6 +32,5 @@ seqan3_test(single_pass_input_test.cpp)
 seqan3_test(get_test.cpp)
 seqan3_test(kmer_hash_test.cpp)
 seqan3_test(interleave_test.cpp)
-seqan3_test(minimiser_test.cpp)
 seqan3_test(minimiser_hash_test.cpp)
 seqan3_test(zip_test.cpp)
