+++ test/unit/search/views/CMakeLists.txt
@@ -1,0 +1,1 @@
+seqan3_test(minimiser_test.cpp)
