+++ test/unit/search/views/minimiser_test.cpp
@@ -15,13 +15,13 @@
 #include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/drop.hpp>
 #include <seqan3/range/views/kmer_hash.hpp>
-#include <seqan3/range/views/minimiser.hpp>
 #include <seqan3/range/views/take_until.hpp>
+#include <seqan3/search/views/minimiser.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
 
 #include <gtest/gtest.h>
 
-#include "../iterator_test_template.hpp"
+#include "../../range/iterator_test_template.hpp"
 
 using seqan3::operator""_dna4;
 using seqan3::operator""_shape;
